removed all 64bit times, reverted to Nordic vanilla

+++ include/third/protocol/trickle.h
@@ -49,7 +49,7 @@ SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 */
 typedef struct
 {
-    uint64_t        t;              /* Absolute value of t. Equals g_trickle_time (at set time) + t_relative */
+    uint32_t        t;              /* Absolute value of t. Equals g_trickle_time (at set time) + t_relative */
     uint32_t        volatile i;              /* Absolute value of i. Equals g_trickle_time (at set time) + i_relative */
     uint32_t        volatile i_relative;     /* Relative value of i. Represents the actual i value in IETF RFC6206 */
     uint8_t         c;              /* Consistent messages counter */
@@ -71,7 +71,7 @@ void trickle_time_increment(void);
 /**
 * @brief set absolute value of global trickle timestamp
 */
-void trickle_time_update(uint64_t time);
+void trickle_time_update(uint32_t time);
 
 /**
 * @brief initialize a trickle algorithm instance. Prepares all flags and 
@@ -113,11 +113,11 @@ void trickle_step(trickle_t* trickle, bool* out_do_tx);
 /**
 * @brief get current global trickle timestamp
 */
-uint64_t trickle_timestamp_get(void);
+uint32_t trickle_timestamp_get(void);
 
 /**
 * @brief get the next time the indicated trickle instance is required to do some processing
 */
-uint64_t trickle_next_processing_get(trickle_t* trickle);
+uint32_t trickle_next_processing_get(trickle_t* trickle);
 
 #endif /* _TRICKLE_H__ */
